+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Always serialize nodes with an expression or a link, even if they have the default value. #585
 - Support cloned group nodes or hard links between groups. #568 #579 #594 #598
 - Default keyframe interpolation method for strokes and shapes is now "Smooth" (was "Linear"). #597
+- Fix animation of Roto Beziers with non-broken tangents. #102
 
 ### Plugins